challenges/misc/table/requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
numpy==1.19.4
periodictable==1.5.3
Pillow==8.0.1
pyparsing==2.4.7
requests==2.25.0
stepic==0.5.0
urllib3==1.26.2
